Rename storage types (#262)
