+++ packages/storage-documents/src/index.ts
@@ -3,7 +3,7 @@
 export { PhysicalLocation, Location, RuleResult, Product, IssueScanResult, IssueScanResults } from './issue-scan-result';
 export { ItemType } from './item-type';
 export { PageIssueScanResult, PageCrawlResult, PageIssueScanRunResult, PageCrawlRunResult, PageScanResult } from './page-scan-result';
-export * from './on-demand-page-scan';
+export * from './on-demand-page-scan-result';
 export { ScanLevel, RunState, ResultLevel, WebsiteScanState } from './states';
 export { WebsitePage, WebsitePageBase, WebsitePageExtra } from './website-page';
 export { PageLastScanResult, Website } from './website';
