+++ packages/storage-documents/src/on-demand-page-scan-result.ts
@@ -9,7 +9,7 @@ export declare type OnDemandPageScanRunState = 'unknown' | 'accepted' | 'queued'
 /**
  * The web page scan run result document.
  */
-export interface OnDemandPageScan extends StorageDocument {
+export interface OnDemandPageScanResult extends StorageDocument {
     scanId: string;
     url: string;
     scanResult?: OnDemandScanResult;
